Create core.magics.extension according to new API.
